requirements-generator.txt
faker==18.13.0
kafka-python==2.0.2
mysql-connector-python==8.0.33
psycopg2-binary==2.9.9
pydantic==2.0.0
python-dotenv==1.0.0
requests==2.31.0
